+++ _src/core/constants.js
@@ -2,8 +2,8 @@ import db from '#db';
 import * as os from 'os';
 
 const CONSTANTS = {
-    FEWU_RELEASE_VERSION: "1.1.2",
-    FEWU_INSIDER_VERSION: "2.4.2",
+    FEWU_RELEASE_VERSION: "1.2.1",
+    // FEWU_INSIDER_VERSION: "2.4.2",
     FEWU_SYMBOLS: ["AR", "NPM", "GH"],
     NODE_CORE_VERSION: process.versions.node,
     NODE_V8_VERSION: process.versions.v8,